✅ update + delete post & user

+++ src/modules/post/post.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
+
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PrismaService } from 'src/modules/prisma/prisma.service';
@@ -25,14 +26,50 @@ export class PostService {
     });
   }
 
-  update(id: number, updatePostDto: UpdatePostDto) {
+  async update(userId: string, id: number, updatePostDto: UpdatePostDto) {
+    const user = await this.prisma.user.findUnique({
+      where: { id: userId },
+    });
+
+    if (!user) throw new NotFoundException('User not found');
+
+    const post = await this.prisma.post.findUnique({
+      where: { id },
+    });
+
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+
+    if (post.authorId !== userId && !user.admin) {
+      throw new ForbiddenException("You can't update this post");
+    }
+
     return this.prisma.post.update({
       where: { id },
       data: updatePostDto,
     });
   }
 
-  remove(id: number) {
+  async remove(userId: string, id: number) {
+    const user = await this.prisma.user.findUnique({
+      where: { id: userId },
+    });
+
+    if (!user) throw new NotFoundException('User not found');
+
+    const post = await this.prisma.post.findUnique({
+      where: { id },
+    });
+
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+
+    if (post.authorId !== userId && !user.admin) {
+      throw new ForbiddenException("You can't delete this post");
+    }
+
     return this.prisma.post.delete({
       where: { id },
     });
