+++ src/modules/user/user.service.ts
@@ -1,14 +1,13 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { User, Prisma } from '@prisma/client';
 import { PrismaService } from 'src/modules/prisma/prisma.service';
+import { UpdateUserDto } from './dto/update-user.dto';
 
 @Injectable()
 export class UserService {
   constructor(private prisma: PrismaService) {}
 
-  async user(
-    userWhereUniqueInput: Prisma.UserWhereUniqueInput,
-  ): Promise<User | null> {
+  async user(userWhereUniqueInput: Prisma.UserWhereUniqueInput): Promise<User | null> {
     return this.prisma.user.findUnique({
       where: userWhereUniqueInput,
     });
@@ -31,20 +30,42 @@ export class UserService {
     });
   }
 
-  async updateUser(params: {
-    where: Prisma.UserWhereUniqueInput;
-    data: Prisma.UserUpdateInput;
-  }): Promise<User> {
-    const { where, data } = params;
+  async updateUser(sessionUserId: string, id: string, data: UpdateUserDto): Promise<User> {
+    const user = await this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) throw new NotFoundException('User not found');
+
+    const sessionUser = await this.prisma.user.findUnique({
+      where: { id: sessionUserId },
+    });
+
+    if (sessionUserId !== id && !sessionUser.admin)
+      throw new ForbiddenException('You can not update this user');
+
     return this.prisma.user.update({
+      where: { id: sessionUserId },
       data,
-      where,
     });
   }
 
-  async deleteUser(where: Prisma.UserWhereUniqueInput): Promise<User> {
+  async deleteUser(sessionUserId: string, id: string): Promise<User> {
+    const user = await this.prisma.user.findUnique({
+      where: { id },
+    });
+
+    if (!user) throw new NotFoundException('User not found');
+
+    const sessionUser = await this.prisma.user.findUnique({
+      where: { id: sessionUserId },
+    });
+
+    if (sessionUserId !== id && !sessionUser.admin)
+      throw new ForbiddenException('You can not delete this user');
+
     return this.prisma.user.delete({
-      where,
+      where: { id },
     });
   }
 }
